Aplicacion en la segunda version

+++ app/src/main/java/com/example/alimentate/data/SensorReading.kt
@@ -5,6 +5,6 @@ package com.example.alimentate.data
 data class SensorReading(
     val tempC: Float,
     val humiPct: Float,
-    val gasRaw: Int,
+    val gasRaw: Float,
     val tsUnix: Long
 )
